requirements.txt
aiomysql==0.0.20
anyio==3.6.2
astroid==2.3.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
azure-core==1.26.4
azure-storage-blob==12.16.0
boto3==1.17.27
botocore==1.20.112
Brotli==1.0.9
cached-property==1.5.1
certifi==2019.11.28
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
ConfigArgParse==1.5.3
cryptography==39.0.2
decorator==5.1.1
dnspython==2.3.0
docutils==0.15.2
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.94.1
fastapi-jwt-auth==0.5.0
Flask==2.2.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.9
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.9.2
httpx==0.23.3
idna==3.4
importlib-metadata==1.2.0
iniconfig==2.0.0
isodate==0.6.1
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
jsonpath-ng==1.5.2
jsonpointer==1.10
jsonschema==3.2.0
lazy-object-proxy==1.4.3
locust==2.15.1
MarkupSafe==2.1.2
mccabe==0.6.1
more-itertools==8.0.2
msgpack==1.0.5
mysql-connector-python==8.0.26
nameparser==1.0.5
oauthlib==2.0.6
packaging==23.0
pathtools==0.1.2
peewee==3.15.4
pluggy==0.13.1
ply==3.11
protobuf==4.22.1
psutil==5.9.4
py==1.11.0
pycparser==2.21
pydantic==1.10.7
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.2
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.1
python-dotenv==1.0.0
python-magic==0.4.15
python-multipart==0.0.5
python-slugify==4.0.1
pyzmq==25.0.2
redis==4.5.4
requests==2.28.2
rfc3986==1.5.0
roundrobin==0.0.4
s3transfer==0.3.7
sentry-sdk==1.19.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
Werkzeug==2.2.3
wrapt==1.11.2
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
